refactor: enhance payroll step progress component and introduce shared dialog components
- Refactored the PayrollStepProgress component to utilize new shared dialog components for settling and reverting payrolls, improving code reusability and organization.
- Removed obsolete dialog and badge components to streamline the codebase and reduce complexity.
- Introduced new PayrollRevertDialog and PayrollSettleDialog components to encapsulate dialog logic and improve maintainability.
- Added a RevertPreviewTable component to display detailed information for revert actions, enhancing user experience and clarity.
- Updated import paths and cleaned up unused code to ensure consistency across the payroll module.
