+++ app/dashboard/worker/worker-form.tsx
@@ -2,210 +2,32 @@
 
 import * as React from "react";
 import { useRouter } from "next/navigation";
-import {
-    Controller,
-    useForm,
-    useWatch,
-    type Control,
-    type FieldPath,
-    type FieldValues,
-} from "react-hook-form";
+import { Controller, useForm, useWatch } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
-import {
-    Banknote,
-    Briefcase,
-    Building2,
-    Clock,
-    CreditCard,
-    Globe,
-    Mail,
-    Phone,
-    User,
-    Users,
-} from "lucide-react";
-import { Button } from "@/components/ui/button";
-import {
-    InputGroup,
-    InputGroupAddon,
-    InputGroupInput,
-} from "@/components/ui/input-group";
-import {
-    Select,
-    SelectContent,
-    SelectItem,
-    SelectTrigger,
-    SelectValue,
-} from "@/components/ui/select";
-import {
-    Card,
-    CardContent,
-    CardDescription,
-    CardHeader,
-} from "@/components/ui/card";
-import {
-    Field,
-    FieldError,
-    FieldGroup,
-    FieldLabel,
-} from "@/components/ui/field";
-
-function RequiredMark() {
-    return (
-        <span className="text-destructive ml-0.5" aria-hidden="true">
-            *
-        </span>
-    );
-}
 import {
     workerUpsertSchema,
     type WorkerUpsertFormInput,
     type WorkerUpsertValues,
 } from "@/db/schemas/worker-employment";
 import type { WorkerWithEmployment } from "@/db/tables/workerTable";
+import { Button } from "@/components/ui/button";
+import {
+    Card,
+    CardContent,
+    CardDescription,
+    CardHeader,
+} from "@/components/ui/card";
+import { FieldGroup } from "@/components/ui/field";
 import { cn } from "@/lib/utils";
-import { WORKER_SHIFT_PATTERNS } from "@/types/status";
 import { formatEnGbDmyNumericFromCalendar } from "@/utils/time/intl-en-gb";
 import { createWorker, updateWorker } from "./actions";
-import { SegmentedTwoChoiceField } from "./segmented-two-choice-field";
-
-const EMPLOYMENT_TYPE_SEGMENT_OPTIONS = [
-    {
-        value: "Full Time" as const,
-        label: "Full Time",
-        activeClassName:
-            "border-emerald-500 bg-emerald-50 text-emerald-800 dark:bg-emerald-500/20 dark:text-emerald-300 dark:border-emerald-500/50",
-        inactiveHoverClassName:
-            "hover:border-emerald-300 hover:bg-emerald-50/50 dark:hover:border-emerald-500/30 dark:hover:bg-emerald-500/10",
-    },
-    {
-        value: "Part Time" as const,
-        label: "Part Time",
-        activeClassName:
-            "border-sky-500 bg-sky-50 text-sky-800 dark:bg-sky-500/20 dark:text-sky-300 dark:border-sky-500/50",
-        inactiveHoverClassName:
-            "hover:border-sky-300 hover:bg-sky-50/50 dark:hover:border-sky-500/30 dark:hover:bg-sky-500/10",
-    },
-] as const;
-
-const EMPLOYMENT_ARRANGEMENT_SEGMENT_OPTIONS = [
-    {
-        value: "Foreign Worker" as const,
-        label: "Foreign Worker",
-        activeClassName:
-            "border-blue-500 bg-blue-50 text-blue-800 dark:bg-blue-500/20 dark:text-blue-300 dark:border-blue-500/50",
-        inactiveHoverClassName:
-            "hover:border-blue-300 hover:bg-blue-50/50 dark:hover:border-blue-500/30 dark:hover:bg-blue-500/10",
-    },
-    {
-        value: "Local Worker" as const,
-        label: "Local Worker",
-        activeClassName:
-            "border-amber-500 bg-amber-50 text-amber-800 dark:bg-amber-500/20 dark:text-amber-300 dark:border-amber-500/50",
-        inactiveHoverClassName:
-            "hover:border-amber-300 hover:bg-amber-50/50 dark:hover:border-amber-500/30 dark:hover:bg-amber-500/10",
-    },
-] as const;
-
-const SHIFT_PATTERN_SEGMENT_OPTIONS = [
-    {
-        value: WORKER_SHIFT_PATTERNS[0],
-        label: WORKER_SHIFT_PATTERNS[0],
-        activeClassName:
-            "border-orange-400 bg-orange-50 text-orange-900 dark:bg-orange-500/20 dark:text-orange-200 dark:border-orange-400/70",
-        inactiveHoverClassName:
-            "hover:border-orange-300 hover:bg-orange-50/50 dark:hover:border-orange-500/35 dark:hover:bg-orange-500/10",
-    },
-    {
-        value: WORKER_SHIFT_PATTERNS[1],
-        label: WORKER_SHIFT_PATTERNS[1],
-        activeClassName:
-            "border-indigo-500 bg-indigo-50 text-indigo-900 dark:bg-indigo-500/25 dark:text-indigo-200 dark:border-indigo-500/55",
-        inactiveHoverClassName:
-            "hover:border-indigo-300 hover:bg-indigo-50/60 dark:hover:border-indigo-500/35 dark:hover:bg-indigo-500/12",
-    },
-] as const;
+import { EmploymentTermsFields } from "./_shared/employment-terms-fields";
+import { PaymentMethodFields } from "./_shared/payment-method-fields";
+import { WorkerIdentityFields } from "./_shared/worker-identity-fields";
 
 function optionalNumberString(n: number | null | undefined): string {
     if (n == null || Number.isNaN(n)) return "";
     return String(n);
-}
-
-function bindTextNumericField(field: {
-    value: unknown;
-    onChange: (v: string) => void;
-    onBlur: () => void;
-    name: string;
-    ref: React.Ref<HTMLInputElement>;
-}) {
-    const str =
-        field.value == null || field.value === "" ? "" : String(field.value);
-    return {
-        name: field.name,
-        ref: field.ref,
-        onBlur: field.onBlur,
-        value: str,
-        onChange: (e: React.ChangeEvent<HTMLInputElement>) => {
-            field.onChange(e.target.value);
-        },
-    };
-}
-
-function NumericControllerField<T extends FieldValues>({
-    control,
-    name,
-    formId,
-    label,
-    icon: Icon,
-    required = false,
-    inputMode = "decimal",
-    visible = true,
-    disabled = false,
-}: {
-    control: Control<T>;
-    name: FieldPath<T>;
-    formId: string;
-    label: string;
-    icon: React.ComponentType<{ className?: string }>;
-    required?: boolean;
-    inputMode?: "decimal" | "numeric";
-    visible?: boolean;
-    disabled?: boolean;
-}) {
-    if (!visible) return null;
-
-    const id = `${formId}-${String(name)}`;
-
-    return (
-        <Controller
-            name={name}
-            control={control}
-            render={({ field, fieldState }) => (
-                <Field data-invalid={fieldState.invalid} className="space-y-2">
-                    <FieldLabel htmlFor={id}>
-                        {label}
-                        {required ? <RequiredMark /> : null}
-                    </FieldLabel>
-                    <InputGroup>
-                        <InputGroupInput
-                            {...bindTextNumericField(field)}
-                            id={id}
-                            type="text"
-                            inputMode={inputMode}
-                            aria-invalid={fieldState.invalid}
-                            aria-required={required ? true : undefined}
-                            disabled={disabled}
-                        />
-                        <InputGroupAddon>
-                            <Icon className="size-4 text-muted-foreground" />
-                        </InputGroupAddon>
-                    </InputGroup>
-                    {fieldState.invalid && (
-                        <FieldError errors={[fieldState.error]} />
-                    )}
-                </Field>
-            )}
-        />
-    );
 }
 
 type WorkerFormValues = WorkerUpsertFormInput;
@@ -296,6 +118,7 @@ export function WorkerForm({ worker, disabled = false }: WorkerFormProps) {
     const paymentMethod = form.watch("paymentMethod");
 
     const isFullTime = employmentType === "Full Time";
+    const isLocalWorker = employmentArrangement === "Local Worker";
 
     return (
         <Card>
@@ -385,441 +208,25 @@ export function WorkerForm({ worker, disabled = false }: WorkerFormProps) {
                     className="space-y-6"
                     autoComplete="off">
                     <FieldGroup className="space-y-6">
-                        <div
-                            className={cn(
-                                "grid gap-4",
-                                isCreate ? "md:grid-cols-2" : "md:grid-cols-2",
-                            )}>
-                            <Controller
-                                name="name"
-                                control={form.control}
-                                render={({ field, fieldState }) => (
-                                    <Field
-                                        data-invalid={fieldState.invalid}
-                                        className="space-y-2">
-                                        <FieldLabel htmlFor={`${formId}-name`}>
-                                            Name
-                                            <RequiredMark />
-                                        </FieldLabel>
-                                        <InputGroup>
-                                            <InputGroupInput
-                                                {...field}
-                                                id={`${formId}-name`}
-                                                aria-invalid={
-                                                    fieldState.invalid
-                                                }
-                                                aria-required
-                                                disabled={disabled}
-                                            />
-                                            <InputGroupAddon>
-                                                <User className="size-4 text-muted-foreground" />
-                                            </InputGroupAddon>
-                                        </InputGroup>
-                                        {fieldState.invalid && (
-                                            <FieldError
-                                                errors={[fieldState.error]}
-                                            />
-                                        )}
-                                    </Field>
-                                )}
-                            />
-                        </div>
-
-                        <div className="grid gap-4 md:grid-cols-2">
-                            <Controller
-                                name="email"
-                                control={form.control}
-                                render={({ field, fieldState }) => (
-                                    <Field
-                                        data-invalid={fieldState.invalid}
-                                        className="space-y-2">
-                                        <FieldLabel htmlFor={`${formId}-email`}>
-                                            Email
-                                        </FieldLabel>
-                                        <InputGroup>
-                                            <InputGroupInput
-                                                {...field}
-                                                value={field.value ?? ""}
-                                                id={`${formId}-email`}
-                                                type="email"
-                                                aria-invalid={
-                                                    fieldState.invalid
-                                                }
-                                                disabled={disabled}
-                                            />
-                                            <InputGroupAddon>
-                                                <Mail className="size-4 text-muted-foreground" />
-                                            </InputGroupAddon>
-                                        </InputGroup>
-                                        {fieldState.invalid && (
-                                            <FieldError
-                                                errors={[fieldState.error]}
-                                            />
-                                        )}
-                                    </Field>
-                                )}
-                            />
-                            <Controller
-                                name="phone"
-                                control={form.control}
-                                render={({ field, fieldState }) => (
-                                    <Field
-                                        data-invalid={fieldState.invalid}
-                                        className="space-y-2">
-                                        <FieldLabel htmlFor={`${formId}-phone`}>
-                                            Phone
-                                        </FieldLabel>
-                                        <InputGroup>
-                                            <InputGroupInput
-                                                {...field}
-                                                value={field.value ?? ""}
-                                                id={`${formId}-phone`}
-                                                aria-invalid={
-                                                    fieldState.invalid
-                                                }
-                                                disabled={disabled}
-                                            />
-                                            <InputGroupAddon>
-                                                <Phone className="size-4 text-muted-foreground" />
-                                            </InputGroupAddon>
-                                        </InputGroup>
-                                        {fieldState.invalid && (
-                                            <FieldError
-                                                errors={[fieldState.error]}
-                                            />
-                                        )}
-                                    </Field>
-                                )}
-                            />
-                        </div>
-
-                        <div className="grid gap-4 md:grid-cols-2">
-                            <Controller
-                                name="countryOfOrigin"
-                                control={form.control}
-                                render={({ field, fieldState }) => (
-                                    <Field
-                                        data-invalid={fieldState.invalid}
-                                        className="space-y-2">
-                                        <FieldLabel
-                                            htmlFor={`${formId}-countryOfOrigin`}>
-                                            Country of Origin
-                                        </FieldLabel>
-                                        <InputGroup>
-                                            <InputGroupInput
-                                                {...field}
-                                                value={field.value ?? ""}
-                                                id={`${formId}-countryOfOrigin`}
-                                                aria-invalid={
-                                                    fieldState.invalid
-                                                }
-                                                disabled={disabled}
-                                            />
-                                            <InputGroupAddon>
-                                                <Globe className="size-4 text-muted-foreground" />
-                                            </InputGroupAddon>
-                                        </InputGroup>
-                                        {fieldState.invalid && (
-                                            <FieldError
-                                                errors={[fieldState.error]}
-                                            />
-                                        )}
-                                    </Field>
-                                )}
-                            />
-                            <Controller
-                                name="race"
-                                control={form.control}
-                                render={({ field, fieldState }) => (
-                                    <Field
-                                        data-invalid={fieldState.invalid}
-                                        className="space-y-2">
-                                        <FieldLabel htmlFor={`${formId}-race`}>
-                                            Race
-                                        </FieldLabel>
-                                        <InputGroup>
-                                            <InputGroupInput
-                                                {...field}
-                                                value={field.value ?? ""}
-                                                id={`${formId}-race`}
-                                                aria-invalid={
-                                                    fieldState.invalid
-                                                }
-                                                disabled={disabled}
-                                            />
-                                            <InputGroupAddon>
-                                                <Users className="size-4 text-muted-foreground" />
-                                            </InputGroupAddon>
-                                        </InputGroup>
-                                        {fieldState.invalid && (
-                                            <FieldError
-                                                errors={[fieldState.error]}
-                                            />
-                                        )}
-                                    </Field>
-                                )}
-                            />
-                        </div>
-
-                        <div className="grid gap-4 md:grid-cols-2">
-                            <SegmentedTwoChoiceField
-                                control={form.control}
-                                name="employmentType"
-                                label="Employment Type"
-                                icon={<Briefcase className="size-4" />}
-                                disabled={disabled}
-                                options={EMPLOYMENT_TYPE_SEGMENT_OPTIONS}
-                                ariaLabel="Employment type"
-                            />
-                            <SegmentedTwoChoiceField
-                                control={form.control}
-                                name="employmentArrangement"
-                                label="Employment Arrangement"
-                                icon={<Users className="size-4" />}
-                                disabled={disabled}
-                                options={EMPLOYMENT_ARRANGEMENT_SEGMENT_OPTIONS}
-                                ariaLabel="Employment arrangement"
-                            />
-                        </div>
-
-                        <SegmentedTwoChoiceField
+                        <WorkerIdentityFields
                             control={form.control}
-                            name="shiftPattern"
-                            label="Shift Pattern"
-                            icon={<Clock className="size-4" />}
+                            formId={formId}
                             disabled={disabled}
-                            options={SHIFT_PATTERN_SEGMENT_OPTIONS}
-                            ariaLabel="Shift pattern"
                         />
-
-                        <div className="grid gap-4 md:grid-cols-5">
-                            <NumericControllerField
-                                control={form.control}
-                                name="monthlyPay"
-                                formId={formId}
-                                label="Monthly Pay"
-                                icon={Banknote}
-                                required
-                                inputMode="decimal"
-                                visible={isFullTime}
-                                disabled={disabled}
-                            />
-                            <NumericControllerField
-                                control={form.control}
-                                name="hourlyRate"
-                                formId={formId}
-                                label="Hourly Rate"
-                                icon={Banknote}
-                                required
-                                inputMode="decimal"
-                                disabled={disabled}
-                            />
-                            <NumericControllerField
-                                control={form.control}
-                                name="restDayRate"
-                                formId={formId}
-                                label="Rest Day Rate"
-                                icon={Banknote}
-                                required
-                                inputMode="decimal"
-                                visible={isFullTime}
-                                disabled={disabled}
-                            />
-                            <NumericControllerField
-                                control={form.control}
-                                name="minimumWorkingHours"
-                                formId={formId}
-                                label="Minimum Working Hours"
-                                icon={Clock}
-                                inputMode="numeric"
-                                visible={isFullTime}
-                                disabled={disabled}
-                            />
-                            <NumericControllerField
-                                control={form.control}
-                                name="cpf"
-                                formId={formId}
-                                label="CPF"
-                                icon={Banknote}
-                                inputMode="decimal"
-                                visible={
-                                    employmentArrangement === "Local Worker"
-                                }
-                                disabled={disabled}
-                            />
-                        </div>
-
-                        <div className="grid gap-4 md:grid-cols-2">
-                            <Controller
-                                name="paymentMethod"
-                                control={form.control}
-                                render={({ field, fieldState }) => (
-                                    <Field
-                                        data-invalid={fieldState.invalid}
-                                        className="space-y-2">
-                                        <FieldLabel
-                                            htmlFor={`${formId}-paymentMethod`}>
-                                            <span className="flex items-center gap-2">
-                                                <CreditCard className="size-4" />
-                                                Payment Method
-                                            </span>
-                                        </FieldLabel>
-                                        <Select
-                                            value={field.value ?? undefined}
-                                            onValueChange={(val) => {
-                                                const method =
-                                                    val as WorkerFormValues["paymentMethod"];
-                                                field.onChange(method);
-
-                                                if (method === "PayNow") {
-                                                    const currentPhone =
-                                                        form.getValues("phone");
-                                                    const currentPayNowPhone =
-                                                        form.getValues(
-                                                            "payNowPhone",
-                                                        );
-                                                    const phoneTrimmed = String(
-                                                        currentPhone ?? "",
-                                                    ).trim();
-                                                    if (
-                                                        !currentPayNowPhone &&
-                                                        phoneTrimmed
-                                                    ) {
-                                                        form.setValue(
-                                                            "payNowPhone",
-                                                            phoneTrimmed,
-                                                            {
-                                                                shouldDirty: true,
-                                                            },
-                                                        );
-                                                    }
-                                                }
-
-                                                if (method === "Cash") {
-                                                    form.setValue(
-                                                        "payNowPhone",
-                                                        "",
-                                                    );
-                                                    form.setValue(
-                                                        "bankAccountNumber",
-                                                        "",
-                                                    );
-                                                }
-                                            }}
-                                            disabled={disabled}>
-                                            <SelectTrigger
-                                                id={`${formId}-paymentMethod`}
-                                                aria-invalid={
-                                                    fieldState.invalid
-                                                }>
-                                                <SelectValue placeholder="Select payment method" />
-                                            </SelectTrigger>
-                                            <SelectContent>
-                                                <SelectItem value="PayNow">
-                                                    PayNow
-                                                </SelectItem>
-                                                <SelectItem value="Bank Transfer">
-                                                    Bank Transfer
-                                                </SelectItem>
-                                                <SelectItem value="Cash">
-                                                    Cash
-                                                </SelectItem>
-                                            </SelectContent>
-                                        </Select>
-                                        {fieldState.invalid && (
-                                            <FieldError
-                                                errors={[fieldState.error]}
-                                            />
-                                        )}
-                                    </Field>
-                                )}
-                            />
-                            {paymentMethod === "PayNow" && (
-                                <Controller
-                                    name="payNowPhone"
-                                    control={form.control}
-                                    render={({ field, fieldState }) => (
-                                        <Field
-                                            data-invalid={fieldState.invalid}
-                                            className="space-y-2">
-                                            <FieldLabel
-                                                htmlFor={`${formId}-payNowPhone`}>
-                                                PayNow
-                                                <RequiredMark />
-                                            </FieldLabel>
-                                            <InputGroup>
-                                                <InputGroupInput
-                                                    {...field}
-                                                    id={`${formId}-payNowPhone`}
-                                                    inputMode="numeric"
-                                                    pattern="[0-9]*"
-                                                    autoComplete="tel"
-                                                    aria-invalid={
-                                                        fieldState.invalid
-                                                    }
-                                                    aria-required
-                                                    disabled={disabled}
-                                                    onChange={(e) => {
-                                                        field.onChange(
-                                                            e.target.value.replace(
-                                                                /\D/g,
-                                                                "",
-                                                            ),
-                                                        );
-                                                    }}
-                                                    value={field.value ?? ""}
-                                                />
-                                                <InputGroupAddon>
-                                                    <Phone className="size-4 text-muted-foreground" />
-                                                </InputGroupAddon>
-                                            </InputGroup>
-                                            {fieldState.invalid && (
-                                                <FieldError
-                                                    errors={[fieldState.error]}
-                                                />
-                                            )}
-                                        </Field>
-                                    )}
-                                />
-                            )}
-                            {paymentMethod === "Bank Transfer" && (
-                                <Controller
-                                    name="bankAccountNumber"
-                                    control={form.control}
-                                    render={({ field, fieldState }) => (
-                                        <Field
-                                            data-invalid={fieldState.invalid}
-                                            className="space-y-2">
-                                            <FieldLabel
-                                                htmlFor={`${formId}-bankAccountNumber`}>
-                                                Bank Account Number
-                                                <RequiredMark />
-                                            </FieldLabel>
-                                            <InputGroup>
-                                                <InputGroupInput
-                                                    {...field}
-                                                    value={field.value ?? ""}
-                                                    id={`${formId}-bankAccountNumber`}
-                                                    aria-invalid={
-                                                        fieldState.invalid
-                                                    }
-                                                    aria-required
-                                                    disabled={disabled}
-                                                />
-                                                <InputGroupAddon>
-                                                    <Building2 className="size-4 text-muted-foreground" />
-                                                </InputGroupAddon>
-                                            </InputGroup>
-                                            {fieldState.invalid && (
-                                                <FieldError
-                                                    errors={[fieldState.error]}
-                                                />
-                                            )}
-                                        </Field>
-                                    )}
-                                />
-                            )}
-                        </div>
+                        <EmploymentTermsFields
+                            control={form.control}
+                            formId={formId}
+                            disabled={disabled}
+                            isFullTime={isFullTime}
+                            isLocalWorker={isLocalWorker}
+                        />
+                        <PaymentMethodFields
+                            control={form.control}
+                            form={form}
+                            formId={formId}
+                            disabled={disabled}
+                            paymentMethod={paymentMethod}
+                        />
                     </FieldGroup>
 
                     {!disabled && (
